src/test/java/ru/job4j/io/AbuseTest.java
package ru.job4j.io;

import static org.junit.Assert.assertThat;
import static org.hamcrest.Matchers.is;

import org.junit.Rule;
import org.junit.Test;
import org.junit.rules.TemporaryFolder;

import java.io.*;
import java.util.List;

public class AbuseTest {
    @Rule
    public TemporaryFolder folder = new TemporaryFolder();

    @Test
    public void drop() throws IOException {
        File source = folder.newFile("source.txt");
        File target = folder.newFile("target.txt");
        try (PrintWriter out = new PrintWriter(source)) {
            out.println("hello java php");
            out.println("foolish job4j dude");
        }
        Abuse.drop(source.getAbsolutePath(), target.getAbsolutePath(), List.of("php", "foolish"));
        StringBuilder rsl = new StringBuilder();
        try (BufferedReader reader = new BufferedReader(new FileReader(target))) {
            reader.lines()
                    .forEach(rsl::append);
        }
        assertThat(rsl.toString(), is("hello java job4j dude "));
    }

}